+++ src/dto/request/models.rs
@@ -1,8 +1,11 @@
+use garde::Validate;
 use serde::{Deserialize, Serialize};
 use utoipa::ToSchema;
 
-#[derive(Debug, Serialize, Deserialize, ToSchema)]
+#[derive(Debug, Serialize, Deserialize, ToSchema, Validate)]
 pub struct CreateModelRequest {
+    #[garde(length(chars, max = 64))]
     pub name: String,
+    #[garde(skip)]
     pub manufacturer_id: Option<i32>,
 }